+++ src/inverse_problem_for_model/specfem_interface/specfem_interface_mod.F90
@@ -660,7 +660,7 @@ contains
              ! size of single record
              b_reclen_field = CUSTOM_REAL * NDIM * NGLLSQUARE * num_abs_boundary_faces
              ! check integer size limit: size of b_reclen_field must fit onto an 4-byte integer
-             if (num_abs_boundary_faces > 2147483646 / (CUSTOM_REAL * NDIM * NGLLSQUARE)) then
+             if (num_abs_boundary_faces > int(2147483646.0 / (CUSTOM_REAL * NDIM * NGLLSQUARE))) then
                 print *,'reclen needed exceeds integer 4-byte limit: ',b_reclen_field
                 print *,'  ',CUSTOM_REAL, NDIM, NGLLSQUARE, num_abs_boundary_faces
                 print *,'bit size Fortran: ',bit_size(b_reclen_field)
@@ -678,7 +678,7 @@ contains
              ! size of single record
              b_reclen_potential = CUSTOM_REAL * NGLLSQUARE * num_abs_boundary_faces
              ! check integer size limit: size of b_reclen_potential must fit onto an 4-byte integer
-             if (num_abs_boundary_faces > 2147483646 / (CUSTOM_REAL * NGLLSQUARE)) then
+             if (num_abs_boundary_faces > int(2147483646.0 / (CUSTOM_REAL * NGLLSQUARE))) then
                 print *,'reclen needed exceeds integer 4-byte limit: ',b_reclen_potential
                 print *,'  ',CUSTOM_REAL, NGLLSQUARE, num_abs_boundary_faces
                 print *,'bit size Fortran: ',bit_size(b_reclen_potential)
@@ -700,7 +700,7 @@ contains
                 ! size of single record
                 b_reclen_field = CUSTOM_REAL * NDIM * NGLLSQUARE * num_abs_boundary_faces
                 ! check integer size limit: size of b_reclen_field must fit onto an 4-byte integer
-                if (num_abs_boundary_faces > 2147483646 / (CUSTOM_REAL * NDIM * NGLLSQUARE)) then
+                if (num_abs_boundary_faces > int(2147483646.0 / (CUSTOM_REAL * NDIM * NGLLSQUARE))) then
                    print *,'reclen needed exceeds integer 4-byte limit: ',b_reclen_field
                    print *,'  ',CUSTOM_REAL, NDIM, NGLLSQUARE, num_abs_boundary_faces
                    print *,'bit size Fortran: ',bit_size(b_reclen_field)
@@ -717,7 +717,7 @@ contains
                 ! size of single record
                 b_reclen_potential = CUSTOM_REAL * NGLLSQUARE * num_abs_boundary_faces
                 ! check integer size limit: size of b_reclen_potential must fit onto an 4-byte integer
-                if (num_abs_boundary_faces > 2147483646 / (CUSTOM_REAL * NGLLSQUARE)) then
+                if (num_abs_boundary_faces > int(2147483646.0 / (CUSTOM_REAL * NGLLSQUARE))) then
                    print *,'reclen needed exceeds integer 4-byte limit: ',b_reclen_potential
                    print *,'  ',CUSTOM_REAL, NGLLSQUARE, num_abs_boundary_faces
                    print *,'bit size Fortran: ',bit_size(b_reclen_potential)
